Introduce WebSocketCore
